refactor: Merge ModeIcon and PurposeIcon into Avatar component
