+++ src/components/MainModeIcon.jsx
@@ -1,13 +1,14 @@
 import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
 
-import ModeIcon from 'src/components/ModeIcon'
+import Avatar from 'src/components/Avatar'
 import { getMainMode } from 'src/lib/trips'
+import { pickModeIcon, modeToColor } from 'src/components/helpers'
 
 const MainModeIcon = ({ trip }) => {
   const mainMode = useMemo(() => getMainMode(trip), [trip])
 
-  return <ModeIcon mode={mainMode} />
+  return <Avatar icon={pickModeIcon(mainMode)} color={modeToColor(mainMode)} />
 }
 
 MainModeIcon.propTypes = {
