+++ src/components/PurposeEditDialog/PurposeEditDialog.jsx
@@ -6,15 +6,18 @@ import NestedSelectModal from 'cozy-ui/transpiled/react/NestedSelect/Modal'
 
 import { purposes } from 'src/components/helpers'
 import { createGeojsonWithModifiedPurpose } from 'src/components/PurposeEditDialog/helpers'
-import PurposeIcon from 'src/components/PurposeIcon'
+import Avatar from 'src/components/Avatar'
 import { useTrip } from 'src/components/Trip/TripProvider'
 import { OTHER_PURPOSE } from 'src/constants/const'
+import { pickPurposeIcon, purposeToColor } from 'src/components/helpers'
 
 const makeOptions = t => {
   const options = purposes.map(purpose => ({
     id: purpose,
     title: t(`trips.purposes.${purpose}`),
-    icon: <PurposeIcon purpose={purpose} />
+    icon: (
+      <Avatar icon={pickPurposeIcon(purpose)} color={purposeToColor(purpose)} />
+    )
   }))
 
   return { children: options }
